moved CalendarScreenState to seperate screenstate package in UI
